refactor: move workshopLabel to utils
