--- apps/src/code-studio/pd/workshop_dashboard/utils/workshopLabel.ts
@@ -1,2 +1,0 @@
-export const workshopLabel = (label: string): string =>
-  `${label.replace(/workshop$/i, '').trim()} Workshop`;
